pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hh-apply"
version = "1.0.0"
description = "CLI-инструмент для автоматических откликов на вакансии hh.ru"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    "patchright>=1.0.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
]

[project.scripts]
hh-apply = "hh_apply.cli:main"

[tool.setuptools.packages.find]
include = ["hh_apply*"]
